Updated fringefit_single
